[#19/FIX] 컨트롤러 분리
