+++ app.js
@@ -31,8 +31,8 @@ app.use((req, res, next) => {
 app.set('view engine', 'ejs');
 app.set('views', path.join(__dirname, 'views'));
 
-const viewRouter = require('./routers/viewRouters');
-app.use('/', viewRouter);
+const userRouters = require('./routers/userRouter');
+app.use('/', userRouters);
 
 const quizRouter = require('./routers/quizRouter');
 app.use('/quiz', quizRouter);
