+++ lib/auth.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const authController = require('../controllers/authcontroller');
+const authController = require('../controllers/auth.controller');
 
 router.get('/verify', authController.showVerifyPage);
 router.get('/register', authController.showRegisterPage);
